remove the old parser
