+++ src/main/java/BlueTurtle/parsers/FindBugsXMLParser.java
@@ -1,11 +1,11 @@
 package BlueTurtle.parsers;
 
 import java.io.File;
-import java.nio.file.Path;
-import java.nio.file.Paths;
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
+import java.util.regex.Matcher;
 
 import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
@@ -15,6 +15,7 @@ import org.w3c.dom.Element;
 import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
 
+import BlueTurtle.finders.ProjectInfoFinder;
 import BlueTurtle.warnings.FindBugsWarning;
 import BlueTurtle.warnings.Warning;
 
@@ -49,29 +50,36 @@ public class FindBugsXMLParser extends XMLParser {
 			Document doc = dBuilder.parse(inputFile);
 
 			// Normalize the elements of the document.
-			doc.getDocumentElement().normalize();
-			
-			// Get the list of file path of the project.
-			NodeList pathsList = doc.getElementsByTagName("Project");
-			
-			NodeList srcList = doc.getElementsByTagName("SrcDir");
-			
-			if (pathsList != null && pathsList.getLength() > 0) {
-				Element pathElement = (Element) pathsList.item(0);
-				srcList = pathElement.getElementsByTagName("SrcDir");
-			}			
+			doc.getDocumentElement().normalize();		
 
 			// Get all list of files where there are warnings.
 			NodeList nList = doc.getElementsByTagName("file");
-			
-			addWarnings(srcList, nList, findBugsWarnings, categoryInfo);
 
+			for (int i = 0; i < nList.getLength(); i++) {
+				// Get the file from the list.
+				Node file = nList.item(i);
+
+				if (file.getNodeType() == Node.ELEMENT_NODE) {
+					// Convert the node to an element.
+					Element fileElement = (Element) file;
+
+					// Get the name of the file where the warning is from.
+					String fileName = fileElement.getAttribute("classname");
+					
+					// Get all the warnings.
+					NodeList warningList = fileElement.getElementsByTagName("BugInstance");
+					
+					addWarnings(fileName, warningList, nList, findBugsWarnings, categoryInfo);
+
+				}
+			}
 		} catch (Exception e) {
 			e.printStackTrace();
 		}
 		return findBugsWarnings;
 	}
 	
+
 	/**
 	 * add individual warning to the warningList.
 	 * 
@@ -82,121 +90,62 @@ public class FindBugsXMLParser extends XMLParser {
 	 * @param categoryInfo is the category information.
 	 * @return a list of FindBugs warnings.
 	 */
-	public List<Warning> addWarnings(NodeList srcList, NodeList nList, List<Warning> findBugsWarnings, HashMap<String, String> categoryInfo) {
+	public List<Warning> addWarnings(String fileName, NodeList warningList, NodeList nList, List<Warning> findBugsWarnings, HashMap<String, String> categoryInfo) {
 		
-		for (int i = 0; i < nList.getLength(); i++) {
-			// Get the file from the list.
-			Node file = nList.item(i);
-			
-			String pathFront = "";
+		for (int j = 0; j < warningList.getLength(); j++) {
+			// Get the warning from the list of warnings.
+			Node warning = warningList.item(j);
 
-
-			if (file.getNodeType() == Node.ELEMENT_NODE) {
+			if (warning.getNodeType() == Node.ELEMENT_NODE) {
 				// Convert the node to an element.
-				Element fileElement = (Element) file;
+				Element warningElement = (Element) warning;
 
-				// Get the class name where the warning is from.
-				String className = fileElement.getAttribute("classname");
-				
-				// split the class name into a string array.
-				String [] classArray = className.split("\\.");
-				
-				// the last one is the class name.
-				className = classArray[classArray.length - 1];
-				
-				// concatenate the source path with the class name.
-				String fileN = className + ".java";
-				
-				fileN = fileN.substring(fileN.lastIndexOf(File.separatorChar) + 1, fileN.length());
+				// message of warning
+				String message = warningElement.getAttribute("message");
 
-				// get the file path from the file name.
-//				String filePath = new File(fileN).getCanonicalPath();
-				Path p = Paths.get(fileN);
-				String filePath = p.toAbsolutePath().toString();
+				// category of warning
+				String category = warningElement.getAttribute("category");
 				
+				// priority of warning
+				String priority = warningElement.getAttribute("priority");
 				
-				/***********************************************************/
-				/*********This part is for get absolute file path***********/
-				
-				// Get the class name where the warning is from.
-				String classN = fileElement.getAttribute("classname");
-				
-				// replace the . with \\ in the file name.
-				String cN = classN.replaceAll("\\.", "\\\\");
-				
-				
-				// initially start with 0
-				int k = 0;
-				
-				// With an empty file path
-				String absoluteFilePath = "";
-				
-				String path = "";
-				// continue go down the list of source path if the file does not exist.
-//				do {
-				
-					pathFront = srcList.item(k).getTextContent();
-	
-					// concatenate the source path with the class name.
-					String fileConcate = pathFront + "\\" + cN + ".java";
+				// line number where the warning is located.
+				int line = Integer.parseInt(warningElement.getAttribute("lineNumber"));
 
-					// get the absoluteFilePath.
-					absoluteFilePath = new File(fileConcate).getAbsolutePath();
+				// Get the category of the warning.
+				String ruleName = warningElement.getAttribute("type");
 				
-//					// increment the counter
-//					k++;
-//					
-//					File f = new File(absoluteFilePath);
-//					path = f.toURI().toURL().getPath();
-//					// check if the file exits or not.
-//				} while(!new File(path).exists());
-				/***********************************************************/
-				/*********This part is for get absolute file path***********/
+				String classification = categoryInfo.get(ruleName);
 				
-				// Get the name of the file where the warning is from.
-				String fileName = filePath.substring(filePath.lastIndexOf(File.separatorChar) + 1, filePath.length());
+				// get the classPaths list from ProjectInfoFinder.
+				ArrayList<String> classPaths = ProjectInfoFinder.getClassPaths();
 				
-				// Get all the warnings.
-				NodeList warningList = fileElement.getElementsByTagName("BugInstance");
-
-				for (int j = 0; j < warningList.getLength(); j++) {
-					// Get the warning from the list of warnings.
-					Node warning = warningList.item(j);
-
-					if (warning.getNodeType() == Node.ELEMENT_NODE) {
-						// Convert the node to an element.
-						Element warningElement = (Element) warning;
-
-						// message of warning
-						String message = warningElement.getAttribute("message");
-
-						// category of warning
-						String category = warningElement.getAttribute("category");
-						
-						// priority of warning
-						String priority = warningElement.getAttribute("priority");
-						
-						// line number where the warning is located.
-						int line = Integer.parseInt(warningElement.getAttribute("lineNumber"));
-
-						// Get the category of the warning.
-						String ruleName = warningElement.getAttribute("type");
-						
-						String classification = categoryInfo.get(ruleName);
-
-						FindBugsWarning fbw = new FindBugsWarning(filePath, fileName, line, message, category, priority, ruleName, classification);
-						
-						fbw.setAbsoluteFilePath(absoluteFilePath);
-						// Add warning to the list of warnings.
-						findBugsWarnings.add(fbw);
+				// replace the dot in the file name with file separator. 
+				String fileNWithSep = fileName.replaceAll("\\.", Matcher.quoteReplacement(File.separator)) + ".java";
+				
+				String filePath = " ";
+				
+				// Loop through all classPathes.
+				for(int i = 0; i < classPaths.size(); i++) {
+					// if the classPath indeed ends with the right class.
+					if (classPaths.get(i).endsWith(fileNWithSep)){
+						// update the file path.
+						filePath = classPaths.get(i);
+						break;
 					}
 				}
+				// Get the name of the file where the warning is from.
+				fileName = fileNWithSep.substring(fileNWithSep.lastIndexOf(File.separatorChar) + 1, fileNWithSep.length());
+				
+				// Construct the new FindBugsWarning.
+				FindBugsWarning fbw = new FindBugsWarning(filePath, fileName, line, message, category, priority, ruleName, classification);
+				
+				// Add warning to the list of warnings.
+				findBugsWarnings.add(fbw);
 			}
 		}
 		return findBugsWarnings;
-		
 	}
-		
 	
 
 }
